requirements.txt
streamlit
google-generativeai
fpdf2
matplotlib
python-dotenv
Pillow
pandas
plotly
gspread
google-auth
openpyxl
requests
